pyproject.toml
[project]
name = "dnd"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
]
